Course Timetabling: Room Ratio  - store room ratio in float precision (as it is done in UniTime) to avoid different number precision issues  - this fixes the issue with meet together constraint failing to assign two classes of 20 (with room ration 0.8) into a room of 32 due to the increased precision in the room ratio    (0.8 float value came as 0.800000011920929 double value, minimal room size computation ceil(room ratio x class limit) was then 33 instead of the expected 32)

+++ trunk/src/net/sf/cpsolver/coursett/TimetableXMLLoader.java
@@ -484,7 +484,7 @@ public class TimetableXMLLoader extends TimetableLoader {
 
             int minClassLimit = 0;
             int maxClassLimit = 0;
-            double room2limitRatio = 1.0;
+            float room2limitRatio = 1.0f;
             if (!"true".equals(classEl.attributeValue("committed"))) {
                 if (classEl.attributeValue("expectedCapacity") != null) {
                     minClassLimit = maxClassLimit = Integer.parseInt(classEl.attributeValue("expectedCapacity"));
@@ -492,7 +492,7 @@ public class TimetableXMLLoader extends TimetableLoader {
                             .attributeValue("expectedCapacity")));
                     if (minClassLimit == 0)
                         minClassLimit = maxClassLimit = roomCapacity;
-                    room2limitRatio = (minClassLimit == 0 ? 1.0 : ((double) roomCapacity) / minClassLimit);
+                    room2limitRatio = (minClassLimit == 0 ? 1.0f : ((float) roomCapacity) / minClassLimit);
                 } else {
                     if (classEl.attribute("classLimit") != null) {
                         minClassLimit = maxClassLimit = Integer.parseInt(classEl.attributeValue("classLimit"));
@@ -500,7 +500,7 @@ public class TimetableXMLLoader extends TimetableLoader {
                         minClassLimit = Integer.parseInt(classEl.attributeValue("minClassLimit"));
                         maxClassLimit = Integer.parseInt(classEl.attributeValue("maxClassLimit"));
                     }
-                    room2limitRatio = Double.parseDouble(classEl.attributeValue("roomToLimitRatio", "1.0"));
+                    room2limitRatio = Float.parseFloat(classEl.attributeValue("roomToLimitRatio", "1.0"));
                 }
             }
